Fixing add/remove buttons and display for meeting sponsor/co-sponsor on edit/add pages
For co-sponsor and second co-sponsor on the meeting add/edit pages, the button coordination was insufficient.  For example: one could add a second co-sponsor even when there wasn't a first co-sponsor.  This was fixed by expanding secondbuttons.js, and modifying person.php to show the appropriate initial states of the divs and buttons.